base_b/requirements.txt
Django==2.1
django-rest-framework==0.1.0
flake8==3.6.0
gunicorn==19.6.0
django-cors-headers==2.4.0